Locking down mutations (users)

+++ graphQL-prisma/src/resolvers/Mutation.js
@@ -45,23 +45,21 @@ const Mutation = {
             token: jwt.sign({userId: user.id}, 'thisisasecret')
         }
     },
-    async deleteUser(parent, args, { prisma }, info) {
-        const userExists = await prisma.exists.User({ id: args.id })
-
-        if(!userExists) {
-            throw new Error('User not found')
-        }
+    async deleteUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
 
         return prisma.mutation.deleteUser({
             where: {
-                id: args.id
+                id: userId
             }
-        }, info)
+        })
     },
-    async updateUser(parent, args, { prisma }, info) {
+    async updateUser(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+
         return prisma.mutation.updateUser({
             where: {
-                id: args.id
+                id: userId
             },
             data: args.data
         }, info)
@@ -82,10 +80,22 @@ const Mutation = {
             }
         }, info)
     },
-    deletePost(parent, args, { prisma }, info) {
+    async deletePost(parent, args, { prisma, request }, info) {
+        const userId = getUserId(request)
+        const postExists =  await prisma.exists.Post({
+            id: args.id,
+            author: {
+                id: userId
+            }
+        })
+
+        if(!postExists) {
+            throw new Error('Unable to delete post')
+        }
+
         return prisma.mutation.deletePost({
             where: {
-                id: args.id
+                id: userId
             }
         }, info)
     },
